+++ examples/fileServer.js
@@ -6,7 +6,7 @@ app.use(mach.logger);
 app.use(mach.modified);
 app.use(mach.file, {
   root: __dirname + '/..',
-  index: 'README.md',
+  autoIndex: true,
   useLastModified: true,
   useETag: true
 });